Extract move interceptor logic into game-specific components.
