+++ src/client/grid/good.module.css
@@ -17,3 +17,7 @@
 .yellow {
   --good-color: #f5ac11;
 }
+
+.white {
+  --good-color: #e9dcc9;
+}
