+++ src/client/index.css
@@ -70,7 +70,8 @@ body.dark-mode .ui.vertical.menu.accordion {
 body.dark-mode .ui.menu.accordion .item {
   color: rgba(255, 255, 255, 0.9);
 }
-body.dark-mode .ui.checkbox label, body.dark-mode .ui.checkbox + label {
+body.dark-mode .ui.checkbox label,
+body.dark-mode .ui.checkbox + label {
   color: rgba(255, 255, 255, 0.9) !important;
 }
 body.dark-mode .ui.form .inline.field > label,
@@ -80,8 +81,13 @@ body.dark-mode .ui.form .inline.fields .field > p,
 body.dark-mode .ui.form .inline.fields > label {
   color: rgba(255, 255, 255, 0.9);
 }
-body.dark-mode .ui.inverted.input > input {
+body.dark-mode .ui.input > input {
   border: none;
+}
+/* This does not apply dark mode to modals, but ensures text is visible on non-inverted modal even in dark mode. */
+body.dark-mode .ui.modal,
+body.dark-mode .ui.modal > .header {
+  color: black;
 }
 
 /**
