+++ src/maps/alabama_railways/view_settings.ts
@@ -2,6 +2,7 @@ import { MapViewSettings } from "../view_settings";
 import { AlabamaRivers } from "./rivers";
 import { AlabamaRailwaysRules } from "./rules";
 import { AlabamaRailwaysMapSettings } from "./settings";
+import { AlabamaRailwaysMoveInterceptorModal } from "./move_interceptor_modal";
 
 export class AlabamaRailwaysViewSettings
   extends AlabamaRailwaysMapSettings
@@ -10,4 +11,6 @@ export class AlabamaRailwaysViewSettings
   getMapRules = AlabamaRailwaysRules;
 
   getTexturesLayer = AlabamaRivers;
+
+  moveInterceptModal = AlabamaRailwaysMoveInterceptorModal;
 }
